[Android] RoundedIconGenerator moved.
Chromium change:

https://chromium.googlesource.com/chromium/src/+/933d2b3b05dad21b8fe319f26657034d667e9c7e

commit 933d2b3b05dad21b8fe319f26657034d667e9c7e
Author: Theresa Wellington <twellington@chromium.org>
Date:   Thu Oct 3 22:50:19 2019 +0000

    Move even more widgets to chrome/browser/ui/widget/android/

     - ArrowBubbleDrawable
     - ImageTextBubble
     - PulseDrawable
     - PulseInterpolator
     - RoundedIconGenerator
     - TextBubble
     - ViewHighlighter
     - ViewHighlighterTest
     - ViewHighlighterTestUtils

    BUG=951455
